backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78

# Scientific computing
numpy==1.24.3

# Machine Learning (for future AI integration)
# tensorflow==2.13.0
# torch==2.0.1
# torchvision==0.15.2
scikit-learn==1.3.2

# HTTP requests
requests==2.31.0
aiofiles==23.2.1

# Data validation
pydantic==2.5.0

# Database (for future use)
sqlalchemy==2.0.23

# Security (for future authentication)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2